Added first cut of the Reaction struct.
I also reorganized the way that the Species class stores its data to add
const-qualified members. I moved the string caster routines into their
own header to accommodate this. This makes the SpeciesLocalParser class
a little simpler and the Species struct a little more complicated, but
overall it is a good trade.

The Reaction struct is messy and will need to be reorganized, so I
entered issue #7 as a reminder to clean it up. I have a few questions
documented in the header for @guidrymwg.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>